GoogleMapとReactのmerge途中、動きません
