+++ front-app/index.html
@@ -14,12 +14,6 @@
 
     <div id="container"></div>
 
-    <div class="clouds"></div>
-    <div class="stars"></div>
-    <div id="map"></div>
-
-    <script async defer src="components/Map.js"></script>
     <script src="build/bundle.js"></script>
-    <script async defer src="https://maps.googleapis.com/maps/api/js?callback=initMap"></script>
   </body>
 </html>
